requirements.txt
streamlit>=1.28
st-gsheets-connection
Authlib>=1.3.2
pandas
plotly
altair
pygwalker
pyyaml
